requirements.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile --index-url=https://pypi.python.org/simple requirements.in
#
--index-url https://pypi.python.org/simple

asgiref==3.5.1
    # via django
django==4.0.4
    # via
    #   -r requirements.in
    #   django-extensions
    #   djangorestframework
django-extensions==3.1.5
    # via -r requirements.in
djangorestframework==3.13.1
    # via -r requirements.in
psycopg2-binary==2.9.3
    # via -r requirements.in
pytz==2022.1
    # via djangorestframework
sqlparse==0.4.2
    # via django
